feat: remove indexer to runtime/
